\name{p3Bound}
\alias{p3Bound}
\title{Checks if constraint on p3 is met?}
\usage{
  p3Bound(mu, A, B, n, fit.control)
}
\arguments{

  \item{mu}{Value of mixture parameter.}

  \item{A}{Value of mixture parameter.}

  \item{B}{Value of mixture parameter.}

  \item{n}{Value of mixture parameter.}

  \item{fit.control}{List of control arguments generated by \code{\link{generateMixtureControl}}}
}
\value{
  The maximal allowed value for \eqn{p_3}.
}
\description{
  Checks if constraint on p3 is met?
}
\author{
  Jonathan Rosenblatt
}

