\name{test.brain.fit}
\docType{data}
\alias{test.brain.fit}
\title{Mixture estimate maps for group-level inference.}
\usage{test.brain.fit}
\description{
  List of \link[tractor.base]{MriImage} objects created using \code{\link{brainMixtureFit}}. Each object contains an estimate map of one of the parameters of the group-level mixture.
}
\format{List}
\references{
Zandbelt, Bram B, Mariet van Buuren, Rene S Kahn, and Matthijs Vink. 
"Reduced Proactive Inhibition in Schizophrenia Is Related to Corticostriatal Dysfunction and Poor Working Memory." 
Biological Psychiatry 70, no. 12 (December 15, 2011): 1151–1158.  \cr
van Buuren, Mariet, Matthijs Vink, Anca E. Rapcencu, and Rene S. Kahn. 
"Exaggerated Brain Activation During Emotion Processing in Unaffected Siblings of Patients with Schizophrenia." 
Biological Psychiatry 70, no. 1 (July 1, 2011): 81–87. \cr
}

